Main

Description

Com esta documentação pretendemos que se torne mais fácil a leitura e o entendimento do nosso código para fazer o jogo Bomberman. Aqui fica um pequeno resumo da lógica usada para fazer o jogo, e mais à frente abordamos todas as funções uma a uma. Esta parte do trabalho teve o objetivo de o tamanha do jogo para que este ocupe menos espaço quando guardado.

Documentation

decode :: String -> [String]

Função que transforma o estado do jogo comprimido novamente no jogo original.

Exemplos de utilização:

>>> decode "9 @ic,~.nb&* 1 1 0 2 10#0 2 1 !"
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 2 10","0 2 1 !"]

decodeLBomb :: String -> [String]

Função que transforma o estado das bombas e dos jogadores do jogo comprimidas novamente nos valores originais.

Exemplos de utilização:

>>> decodeLBomb "* 1 1 0 2 10#0 2 1 !"
["* 1 1 0 2 10","0 2 1 !"]

decodeLRand :: String -> String

Função que transforma os valores do mapa que não são fixo comprimidos nos seu valores originais.

Exemplos de utilização:

>>> decodeLRand "9 @ic,~.nb&"
"            ?+  +  ? ?  ? ?  !  ???? "

decodeUpperL :: String -> String

Função que descomprime as letras maiusculas.

Exemplos de utilização:

>>> decodeUpperL "A@iB,~.nb&"
"aa@ibb,~.nb&"

decodeNum :: String -> String

Função que descomprime os numeros na string comprimida.

Exemplos de utilização:

>>> decodeNum "A@iB3a,~.nb&"
"A@iBaaa,~.nb&"

decodeSimbol :: String -> String

Função que descomprime os simbolos.

Exemplos de utilização:

>>> decodeSimbol "A@iB,~.nb&"
"AabiBaeebbanbfe"

decodePatternA :: Char -> String

Função que descomprime um simbolo em duas letras.

Exemplos de utilização:

>>> decodePatternA '@'
"ab"

decodePatternB :: Char -> String

Função que descomprime um simbolo em duas letras.

Exemplos de utilização:

>>> decodePatternB ':'
"be"

decodePatternE :: Char -> String

Função que descomprime um simbolo em duas letras.

Exemplos de utilização:

>>> decodePatternE '~'
"eb"

decodePatternF :: Char -> String

Função que descomprime um simbolo em duas letras.

Exemplos de utilização:

>>> decodePatternF '$'
"fb"

decodeLowerL :: String -> String

Função que descomprime as letras minusculas.

Exemplos de utilização:

>>> decodeLowerL "aa@ibb,~.nb&"
"    @+  ? ?,~.!  ?&"

decodePattern :: Char -> String

Função que descomprime as letras minusculas.

patternVazio :: Char -> String

Função que descomprime as letras minusculas.

patternPedra :: Char -> String

Função que descomprime as letras minusculas.

patternPlus :: Char -> String

Função que descomprime as letras minusculas.

patternExcl :: Char -> String

Função que descomprime as letras minusculas.

conv :: String -> Int

Função que converte um string num inteiro.

Exemplos de utilização:

>>> conv "120"
120

cString :: Char -> String

Função que converte um string num inteiro.

Exemplos de utilização:

>>> cString '+'
"+"

map1 :: (a -> Int -> Int -> b) -> [a] -> Int -> Int -> [b]

Função que altera um elemento de uma lista.

encode :: [String] -> String

Função que comprime o mapa.

Exemplos de utilização:

>>> encode ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 2 10","0 2 1 !"]
"9 @ic,~.nb&* 1 1 0 2 10#0 2 1 !"

junta :: [String] -> String

Função que comprime a informação dos jogadores.

Exemplos de utilização:

>>> junta ["0 1 1","1 2 2"]
"#0 1 1#1 2 2"

coordPUP :: Char -> [String] -> [(Int, Int)]

Função usada para ir buscar as coordenadas dos pontos se têm power ups de um certo tipo.

Exemplos de utilização:

>>> coordPUP '+' ["+ 5 2","+ 3 3","! 5 5"]
[(5,2),(3,3)]

dec :: String -> (Int, Int)

Função usada para ir buscar as coordenadas de alguma coisa que necessitamos.

Exemplos de utilização:

>>> dec "0 1 1"
(1,1)

>>> dec "1 3 2 ++!"
(3,2)

coloPUPnMplusT :: [String] -> [(Int, Int)] -> [(Int, Int)] -> [String]

Função usada para colocar os powerups no mapa.

coloPUPnMplus :: [String] -> (Int, Int) -> Char -> Char -> [String]

Função usada para colocar um powerup no mapa.

alteraL :: String -> Int -> Char -> Char -> String

Função usada para alterar a informaçao de uma linha do mapa.

>>> alteraL "#  ?     #" 3 '+' '>'
"#  +     #"

newMapWPUP :: [String] -> [String]

Função retira os power da parte de baixo do mapa.

Exemplos de utilização:

>>> newMapWPUP ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]
["#########","#       #","# #?#+# #","#  +  ? #","#?# # #?#","# ?  !  #","# #?#?# #","#  ??   #","#########"]

takeRandValue :: [String] -> Int -> Int -> String

Função que retira os valores aleatórios de um mapa.

>>> takeRandValue ["#########","#       #","# #?#+# #","#  +  ? #","#?# # #?#","# ?  !  #","# #?#?# #","#  ??   #","#########"] 9 0
"   ?+  +  ? ?  ? ?  !  ???? "

randPerLine :: String -> Int -> Int -> String

Função que retira os valores aleatórios de uma linha do mapa.

>>> randPerLine "# #?#+# #" 9 2
"?+"

linha1 :: String -> String

Função que retira os valores aleatórios da primeira ou última linha do mapa.

linha2 :: String -> Int -> Int -> String

Função que retira os valores aleatórios da segunda linha e da penúltima linha do mapa.

>>> linha2 "#       #" 0 9
"   "

linha3 :: String -> Int -> Int -> String

Função que retira os valores aleatórios da terceira e antepenúltima linha do mapa.

>>> linha3 "# #?#+# #" 0 9
"?+"

linha4 :: String -> String

Função que retira os valores aleatórios das restantes linhas do mapa.

nSeguidos :: Char -> String -> Int

Função calcula o número de elementos iguais seguidos sendo que o char inicial pertence à lista.

>>> nSeguidos '#' "#######"
8

nSeguidosSTR2 :: String -> String

Função que se tiver duas letras minusculas iguais seguidas transforma-as numa maiuscula.

>>> nSeguidosSTR2 "aabkbb"
"AbkB"

nSeguidosSTR :: String -> String

Função que transforma um número de carateres seguidos numa string em que a cabeça é o numero de vezes que esse carater aparece seguido.

>>> nSeguidosSTR "ab????a"
"ab4?a"

parteList :: String -> [String]

Função que se tiver duas letras minusculas iguais seguidas transforma-as numa maiuscula.

>>> parteList "3? ?+!?"
["3?"," ?","+!","?"]

aux1 :: String -> String

Função auxiliar da parteList.

aux2 :: String -> Int -> String

Função auxiliar da parteList.

compAlpha :: [String] -> [String]

Função que transforma o resultado de parteList em letras.

>>> compAlpha ["3?"," ?","+!","?"]
["3?","b","m","?"]

subEsp :: String -> String

Função auxiliar da comAlpha se ve o padrão.

subInt :: String -> String

Função auxiliar da comAlpha se ve o padrão.

subPlus :: String -> String

Função auxiliar da comAlpha se ve o padrão.

subEx :: String -> String

Função auxiliar da comAlpha se ve o padrão.

subPLa :: String -> String

Função auxiliar da comBeta se ve o padrão.

subPLb :: String -> String

Função auxiliar da comBeta se ve o padrão.

subPLe :: String -> String

Função auxiliar da comBeta se ve o padrão.

subPLf :: String -> String

Função auxiliar da comBeta se ve o padrão.

compBeta :: [String] -> [String]

Função que transforma duas letras num simbolo.

>>> compBeta ["?a","ab","ba","ik"]
["?a","@",".","ik"]

rand :: Char -> Char

Função criada com o intuito de ser usada para atribuir a cada um dos numeros inteiros que se encontram na lista dos valores aleatórios um Char.

Exemplos de utilização:

>>> rand 83
' '

geraLinha :: Int -> (Int, Int) -> [(Int, Int)]

Dado a dimensão do mapa e o primeiro ponto duma Linha, esta função devolve os pares de coordenadas de todos os pontos duma Linha.

Exemplos de utilização:

>>> geraLinha 5 (0,0)
[(0,0),(1,0),(2,0),(3,0),(4,0)] 

fillMap2 :: Int -> [(Int, Int)] -> String -> String

Esta função preenche uma linha, ponto a ponto, com os seus determinados caracteres.

linha :: (Int, Int) -> Bool

Esta função é usada para testar se um ponto do mapa, excepto os exteriores, terá uma pedra ou nao

mapa1 :: Int -> String -> [String]

Esta função irá gerar a primeira parte do mapa, ou seja, todos os pontos do mapa. Irá fazê-lo usando a função fillMap que, como dito anteriormente, preenche uma e uma só linha, ponto a ponto. Usamos então uma função auxiliar, que use a função fillMap para a primeira linha e depois a invoque recursivamente para preencher as linhas restantes.

Exemplos de utilização:

>>> mapa1 7 0
["#######","#     #","# # # #","# ??  #","# #?# #","#     #","#######"]

alpha :: Int -> Int

Função auxilidar para tornar a função soma menos complexa.

soma :: Int -> Int -> Int

Esta função conta o numero de valores aleatórios utilizados até à linha em questao.

Exemplos de utilização:

>>> soma 9 4
16

geraParesR :: Int -> (Int, Int) -> [(Int, Int)]

Função que gera devolve as coordenadas de todos os pontos em que é preciso utilizar valores aleatórios.

Exemplos de utilização:

>>> geRaParesR 7 (0,0)
[(3,1),(3,2),(1,3),(2,3),(3,3),(4,3),(5,3),(3,4),(3,5)] 

powerUP :: Int -> [(Int, Int)] -> String -> [String]

Função que adiciona, após os mapa, as Strings com a informação acerca da posição do jogador e dos powerups.

Exemplos de utilização:

>>> powerUP 9 (geraParesR 9 (0,0) (valr 9 0)
["+ 5 2","+ 3 3","! 5 5"]

mapa2 :: Int -> String -> [String]

Função principal, que utiliza tanto a função que gera o mapa em si, como a função que gera as linhas com as coordenadas dos power ups.

Exemplos de utilização:

>>> mapa2 9 0
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]

main :: IO ()