Tarefa6_li1g048

Description

Com esta documentação pretendemos que se torne mais fácil a leitura e o entendimento do nosso código para fazer o jogo Bomberman. Aqui fica um pequeno resumo da lógica usada para fazer o jogo, e mais à frente abordamos todas as funções uma a uma. Esta parte do trabalho teve o objetivo de criar um bot ,ou seja, um programa que dado o estado do jogo, tempo e o número de um jogador conseguisse tomar as suas decisões.

Documentation

type Local = (Int, Int)

Coordenadas de um ponto.

bot :: [String] -> Int -> Int -> Maybe Char

Função principal bot que dado o estado atual do jogo e o tempo, toda a decisão por um jogador.

Exemplos de utilização:

>>> bot ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","0 1 1","1 5 1"] 0 100
Just 'R'

vaiBelem :: String -> (Char, Char)

A função vaiBelem é a função responsável por dar os dois elementos diferente pertencentes a um string para movimentar o bot.

Exemplos de utilização:

>>> vaiBelem "UUUL"
('U','L')

invertCaminho :: String -> String

A função invertCaminho é a função responsável passar o string que faça primeiros os movimentos horizontais e depois os verticais numa que faça primeiro os verticais e só depois os horizontais.

Exemplos de utilização:

>>> invertCaminho "UUUL"
"LLLU"

caminho :: Local -> Local -> String

A função caminho é a função responsável por dar uma string com os movimento para o bot se deslocar de um sitio para outro.

Exemplos de utilização:

>>> caminho (1,1) (4,1)
"RRR"

alteralcaminho :: Local -> String -> String -> String

A função alteralcaminho é a função responsável por tornar o caminho mais eficiente e muitas vezes possivel.

Exemplos de utilização:

>>> alteralcaminho (1,1) "RR" "LL"
"RRLL"

caminhof :: Local -> Local -> String

A função caminhof é a função responsável criar o caminho melhor entre duas posições.

Exemplos de utilização:

>>> caminhof (1,1) (3,3)
"RRLL"

objetivo :: Time -> Dim -> (Mapa, [Ipower], [Iplayer]) -> Int -> Local

A função objetivo é a função responsável por estabelecer o objetivo do bot.

Exemplos de utilização:

>>> objetivo 10 9 (["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"],[],["1 2 1","0 1 1"]) 0
(3,4)

nMovimentos :: String -> Local -> Mapa -> Int

A função nMovimentos conta o numero de ticks que o bot demora a chegar a outro sítio

Exemplos de utilização:

>>> nMovimentos "DDRR" (1,1) ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]
13

escolherCam :: String -> String -> Local -> Mapa -> String

A função escolherCam escolhe o caminho entre dois movimentos diferentes.

Exemplos de utilização:

>>> escolherCam "RRDD" "DDRR" (1,1) ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]
"DDRR"

fazObjtivo :: Time -> Dim -> Int -> Mapa -> Maybe Char

A função fazObjtivo decide o que fazer uma vez na posição pretendida.

Exemplos de utilização:

>>> fazObjtivo 10 9 0 ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","0 1 1"]
Nothing

matarJogador :: Int -> [Iplayer] -> Local

A função matarJogador decide que jogador tem que matar conforme o seu identificador e responde as suas coordenadas.

Exemplos de utilização:

>>> matarJogador 2 ["0 1 1","1 3 1","2 2 3", "3 5 5"]
(1,1)

iden :: Int -> [Iplayer] -> (Local, Bool)

A função iden é responsável por dizer se o jogardor esta vivo e responder um par com as suas coordenadas e o ser estado.

Exemplos de utilização:

>>> iden 1 ["0 1 1","1 3 1","2 2 3", "3 5 5"]
((3,1),True)

mataZ :: Int -> [Iplayer] -> Local

A função mataZ tem os alvos se o jogador do bot for o 0.

mata1 :: Int -> [Iplayer] -> Local

A função mataZ tem os alvos se o jogador do bot for o 1.

mata2 :: Int -> [Iplayer] -> Local

A função mataZ tem os alvos se o jogador do bot for o 2.

mata3 :: Int -> [Iplayer] -> Local

A função mataZ tem os alvos se o jogador do bot for o 3.

coordIni :: Int -> Int -> Local

A função coordIni devolve as coordenadas iniciais do bot.

Exemplos de utilização:

>>> coordIni 1 9
(7,1)

movimento :: World -> Int -> Int -> Local -> [Local] -> Maybe Char

A função movimento é a função responsável por levar o bot a executar o seu objetivo.

movimentar :: Mapa -> Char -> Local -> Maybe Char

A função movimentar é a função que movimenta o bom

mov :: Mapa -> Char -> Local -> Maybe Char

A função mov é uma função auxiliar da movimentar.

verRaid :: Ibomb -> Raio

A função verRaid ve o raio de uma bomba.

infoBomb :: [Ibomb] -> [((Int, Int), Int)]

A função inforBomb transforma um lista de bombas numca lista do tipo [(Local,Int)].

Exemplos de utilização:

>>> infoBomb ["* 1 1 0 2 10","* 2 3 1 1 10"]
[((1,1),2),((2,3),1)]

splitexplosionBot :: Epicentro -> Raio -> Mapa -> [Ipower] -> Lexplo

A função splitexplosionBot dado o centro de uma bomba e o seu raid devolver as coordenadas das explusões.

explosaoTotal :: Mapa -> [Ipower] -> Lexplo -> [Raio] -> Lexplo

A função explosaoTotal é responsável por calcular as coordenadas das explusões de todas as bombas no mapa.

explodeSum :: World -> [(Int, Int)]

A função explosaoSum é responsável por calcular as coordenadas das explusões de todas as bombas no mapa.

ativarShield :: World -> [Local] -> Local -> Time -> Int -> Maybe Char

A função ativarShield é a função que toma decisão sobre o que fazer quando o bot esta dentro do raio de uma explusão.

minimumADA :: [(Local, Int)] -> Local

A função minimumADA calcula o par com a minima segunda componemte e devolve a primeira componente desse par.

Exemplos de utilização:

>>> minimumADA [((1,1),2),((2,2),4)]
(1,1)

maximumADA :: [(Local, Int)] -> Local

A função maximumADA calcula o par com a maxima segunda componemte e devolve a primeira componente desse par.

Exemplos de utilização:

>>> maximumADA [((1,1),2),((2,2),4)]
(2,2)

solve :: Mapa -> [Local] -> [Local] -> Local -> Raio -> Maybe Char

A função solve toma a decisão para o bot se desviar da bomba sem a espiral influenciar essa decisão.

solve2 :: Mapa -> [Local] -> [Local] -> [Local] -> Local -> Raio -> Time -> Maybe Char

A função solve2 toma a decisão para o bot se desviar da bomba com a espiral influenciar essa decisão.

verificaTotal :: Mapa -> Local -> Local -> [Local]

A função verificaTotal calcula as coordenadas pelas quais o bot vai passar, juntando todas as suas auxiliares.

lastver :: Mapa -> [Local] -> [Local]

A função lastver pega nas coordenadas originadas pelo caminho e ve até onde o bot pode chegar.

Exemplos de utilização:

>>> lastver ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"] [(2,1),(3,1),(2,1)]
[(2,1),(3,1),(2,1)]

verLinhas :: Local -> Local -> [Local]

A função verLinha calcula a coordenadas dos movimentos entre duas posições

Exemplos de utilização:

>>> verLinhas (1,1) (3,2)
[(2,1),(3,1),(3,2)]

coordCaminho :: Local -> Char -> Local

A função coordCaminho calcula as novas coordenadas após um movimento.

Exemplos de utilização:

>>> coordCaminho (1,1) 'R'
(2,1)

transCam :: Local -> String -> [Local]

A função transCam calcula as coordenadas por onde o bot vai passar dado um caminho.

Exemplos de utilização:

>>> transCam (1,1) "RRL"
[(2,1),(3,1),(2,1)]

maybeToNumber :: Maybe Int -> Int

A função maybeToNumber transforma um Maybe Int num Int

Exemplos de utilização:

>>> maybeToNumber (Just 2)
2