Main

Description

Com esta documentação pretendemos que se torne mais fácil a leitura e o entendimento do nosso código para fazer o jogo Bomberman. Aqui fica um pequeno resumo da lógica usada para fazer o jogo, e mais à frente abordamos todas as funções uma a uma. Esta tarefa é a responsável pela componente gráfica do bot e por isso tem um papel muito importante no jogo sendo que esta componemte é muito importante.

Documentation

type Map = [String]

Lista de Strings do Mapa.

type Poderes = [String]

Lista de Strings de poderes.

type Bombs = [String]

Lista de Strings de bombas.

type Jodadores = [String]

Lista de Strings de Jogadores.

type Estado = (Map, Int, Int, Int, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, (Picture, Picture, Picture, Picture, Picture, Picture, Picture), (Picture, Picture, Picture, Picture, Picture), (Picture, Picture, Picture, Picture, Picture), (Picture, Picture, Picture, Picture, Picture), (Picture, Picture, Picture, Picture, Picture), [String], [String], (Bool, Picture), (Bool, Picture), Int, (Bool, Picture), Int, (Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture, Picture), Int, (Bool, Picture), Picture, (Picture, Picture, Picture, Picture), Map, Int)

Estado de jogo utilizado pela função play.

type Char1 = (Char, Int, Int)

Novo tipo de coordenada com (Char,Int,Int)

estadoInicial :: Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Int -> Estado

A função estado inicial é a função responsável por criar o estado inicial do jogo e recebe como inputs uma seed para o jogo e todos os BMP que foram usados.

stringjogadores :: Int -> Int -> Int -> Jodadores

A função stringjogadores é a funçao que dado um numero de jogado e o tamanho do map responde as strings iniciais desses jogadores. == Exemplos de utilização:

>>> stringjogadores 4 0 9
["0 1 1","1 7 1","2 1 7","3 7 7"]

escrevestr :: Int -> Int -> String

A função auxiliar da stringjogadores que dado o tamnho do mapa e o numero do jogador cria a string inicial desse jogador. == Exemplos de utilização:

>>> escrevestr 0 9
"0 1 1"

retangolo :: Float -> Picture

A função responsável por criar o retângolo onde é mostrado o tempo do jogo.

color1 :: Color

Cor usada para o painel do Tempo.

tampic :: Int -> Float

A função tampic calculo o tamanho maximo que um elemento do mapa pode ocupar. == Exemplos de utilização:

>>> tampic 9
66.0

transformaT :: Map -> Int -> [Char1]

A função transformT é a responsável por transformar as listas de strings do mapa numa nova lista com o tipo (Char,Int,Int). == Exemplos de utilização:

>>> transformaT ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"] 0
[('#',0,0),('#',1,0),('#',2,0),('#',3,0),('#',4,0),('#',5,0),('#',6,0),('#',7,0),('#',8,0),('#',0,1),(' ',1,1),(' ',2,1),(' ',3,1),(' ',4,1),(' ',5,1),(' ',6,1),(' ',7,1),('#',8,1),('#',0,2),(' ',1,2),('#',2,2),('?',3,2),('#',4,2),('?',5,2),('#',6,2),(' ',7,2),('#',8,2),('#',0,3),(' ',1,3),(' ',2,3),('?',3,3),(' ',4,3),(' ',5,3),('?',6,3),(' ',7,3),('#',8,3),('#',0,4),('?',1,4),('#',2,4),(' ',3,4),('#',4,4),(' ',5,4),('#',6,4),('?',7,4),('#',8,4),('#',0,5),(' ',1,5),('?',2,5),(' ',3,5),(' ',4,5),('?',5,5),(' ',6,5),(' ',7,5),('#',8,5),('#',0,6),(' ',1,6),('#',2,6),('?',3,6),('#',4,6),('?',5,6),('#',6,6),(' ',7,6),('#',8,6),('#',0,7),(' ',1,7),(' ',2,7),('?',3,7),('?',4,7),(' ',5,7),(' ',6,7),(' ',7,7),('#',8,7),('#',0,8),('#',1,8),('#',2,8),('#',3,8),('#',4,8),('#',5,8),('#',6,8),('#',7,8),('#',8,8)]

tranforma :: String -> (Int, Int) -> [Char1]

A função tranforma é auxiliar da transforMat que transformar um linha do mapa numa nova lista com o tipo (Char,Int,Int). == Exemplos de utilização:

>>> tranforma "#######" (0,0)
[('#',0,0),('#',1,0),('#',2,0),('#',3,0),('#',4,0),('#',5,0),('#',6,0)]

topicT :: [Char1] -> Picture -> Picture -> Picture -> [(Picture, Int, Int)]

A função topicT é a responvável por tranforma o tipo Char1 em (Picture,Int,Int) conforme o char da primeira compunente.

charTopic :: Char1 -> Picture -> Picture -> Picture -> (Picture, Int, Int)

A função charTopic é a auxiliar da topicT e é a responvável por tranformar um tipo Char1 em (Picture,Int,Int) conforme o char da primeira compunente.

todaPic :: [(Picture, Int, Int)] -> Float -> [Picture]

A função todaPic é a responvável por colocar as figuras na sua devida posição conforme as suas coordenadas

explosionUfla :: Lexplo -> Picture -> Picture -> Picture -> Int -> [(Picture, Int, Int)]

A função explosionUfla é a responvável por transformar as coordenadas das explusões para cima das bombas em figuras.

explosionLfla :: Lexplo -> Picture -> Picture -> [(Picture, Int, Int)]

A função explosionLfla é a responvável por transformar as coordenadas das explusões das bombas em figuras.

splitexplosionad :: Epicentro -> Raio -> Mapa -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Poderes -> [(Picture, Int, Int)]

A função splitexplosionad é a responvável criar a explusão de uma bomba dado o raio e o centro da explusão.

singleflame :: String -> [String] -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Poderes -> [(Picture, Int, Int)]

A função singleflame é a responsável por passar à splitexplosionad toda a informação necessária para criar o efeito da explusão.

flamesbombas :: [String] -> [String] -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Poderes -> [(Picture, Int, Int)]

A função flamesbombas é a responvável criar a explusão todas as bombas.

tranformaBomba :: [String] -> [String] -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Poderes -> [(Picture, Int, Int)]

A função tranformaBomba é a responvável por separar as bombas que vão expludir das que vão ficar paradas.

drawPower :: Map -> [(Int, Int)] -> Picture -> [(Picture, Int, Int)]

A função drawPower é a responvável verficiar se o powerup deve ser representado ou não conforme esta tapado ou não.

drawPlayer :: Jodadores -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Bombs -> [(Picture, Int, Int)]

A função drawPlayer é a responvável transformar as coordenadas de todos os jogadores nos respetivos tipos (Picture,Int,Int).

drawSingle :: Char -> (Int, Int) -> Picture -> Picture -> Picture -> Picture -> (Picture, Int, Int)

A função drawSingle é a responvável transformar as coordenadas de um jogador no respetivo tipo (Picture,Int,Int).

scoreboard :: [String] -> Int -> (Picture, Picture, Picture, Picture) -> [Picture]

A função scoreboard é a responsável por representar as imagens dos jogadores que ainda estão vivos debaixo do tempo do jogo.

dono :: Int -> (Picture, Picture, Picture, Picture) -> Picture

A função dono é a responvável identificar qual a imagem a usar no scoreboard.

desenhaNumero :: [Picture] -> Int -> [Picture]

A função desenhaNumero é a responvável colocar os números do tempo na posição correta para serem representados.

charToNumber :: Char -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture

A função charToNumber é a responvável por transformar um número na sua respetiva imagem.

desenhaTempo :: Estado -> Picture

A função desenhaTempo é a responvável desenhar o número escrito pelo o utilizador no menu de escolher o tamanho do mapa.

desenhaTempo2 :: Estado -> Picture

A função desenhaTempo2 é a responvável desenhar o número escrito pelo o utilizador no menu de escolher o número de jogadores

desenhaMenu1 :: Estado -> Picture

A função desenhaMenu é a responvável desenhar os menus e alterar os estados dos mesmo para o jogo começar.

desenhaEstado :: Estado -> Picture

A função desenhaEstado é a função mais importante desta tarefa e é a responsável por dar vida à parte grafica do jogo.

newImage :: Char -> Picture -> Picture -> Picture -> Picture -> Picture -> Picture

A função newImagem é a responvável alterar a representação do jogador conforme o comando utilizado.

newfram :: Estado -> Char -> Int -> Estado

A função newfram é a responvável alterar a representação de todos os jogadores conforme o comando utilizado.

reageMenu2 :: Event -> Estado -> Estado

A função reageMenu2 é a responsável por fazer com que o menu2 funcione.

alterDim :: Int -> Estado -> Estado

A função alterDim é a responvável alterar a dimensão do mapa no menu2 sempre que utilizador pressiona uma tecla.

movePre :: Estado -> Char -> Int -> Estado

A função movePre é a um função auxiliar da reageEvento e é a responsável pelo movimento dos jogadores

reageMenu :: Event -> Estado -> Estado

A função reageMenu é a função que decide como o jogo deve responder quando o utilizador pressiona um tecla conforme o seu estado atual.

reageEvento :: Event -> Estado -> Estado

A função reageEvento é a responvável por responder ,já no jogo, quando os utilizadores pressionam as teclas

reageTempo :: Float -> Estado -> Estado

A função reageTempo é a responvável pela para passagem do tempo ja quando o jogo foi iniciado.

reageMenu3 :: Event -> Estado -> Estado

A função reageMenu3 é a responsável por fazer com que o menu3 funcione ,ou seja, é responsável por permitir ao utilizador introduzir o número de utilizadores pretendidos.

alteraNum :: Int -> Estado -> Estado

A função alteraNum é a responsável alterar o numero que vai ser exibido no Menu de escolher o número de utilizadores.

reageTempo2 :: Float -> Estado -> Estado

A função reagemTempo2 é a responsável por indicar ao jogo como este deve reagir à passagem do tempo conforme o estado atual.

reageTMenu1 :: Estado -> Estado

A função reageTMenu1 é a pela passagem do tempo no menu1

fr :: Int

fr indica a FrameRate que vai ser utilizada.

dm :: Display

dm indica todas as informações sobre a janela do jogo.

main :: IO ()

A função ´main´ é a responsável por juntar todas as pequenas peças de modo a tornar o jogo jogável.