Main

Description

Com esta documentação pretendemos que se torne mais fácil a leitura e o entendimento do nosso código para fazer o jogo Bomberman. Aqui fica um pequeno resumo da lógica usada para fazer o jogo, e mais à frente abordamos todas as funções uma a uma. Esta parte do trabalho teve o objetivo de realizar movimentos dado o estado do jogo, um jogador e um comando.

Documentation

conv :: String -> Int

Função que converte Strings para Inteiros.

Exemplos de utilização:

>>> conv "120"
120

dec :: String -> (Int, Int)

Função usada para ir buscar as coordenadas de alguma coisa que necessitamos.

Exemplos de utilização:

>>> dec "0 1 1"
(1,1)

>>> dec "1 3 2 ++!"
(3,2)

rdec :: String -> (Int, Int) -> String

Função usada para alterar as coordenadas de uma String.

Exemplos de utilização:

>>> rdec "0 1 1 ++!" (2,1)
"0 2 1 ++!"

dbomb :: String -> Int

Função que dado a string com a informação da bomba devolve o número do jogador detentor da bomba.

Exemplos de utilização:

>>> dbomb "* 1 1 2 1 10"
2

coord :: Char -> [String] -> (Int, Int)

Função usada para ir buscar as coordenadas do ponto onde se encontra o jogador.

coordPUP :: Char -> [String] -> [(Int, Int)]

Função usada para ir buscar as coordenadas dos pontos se têm power ups de um certo tipo.

Exemplos de utilização:

>>> coordPUP '+' ["+ 5 2","+ 3 3","! 5 5"]
[(5,2),(3,3)]

coordBombas :: [String] -> [(Int, Int)]

Função usada para ir buscar as coordenadas de bombas as bombas colocadas no mapa.

Exemplos de utilização:

>>> coordBombas ["* 1 1 0 2 10","* 2 1 2 3 10"]
    [(1,1),(2,1)]

coordCompare :: (Int, Int) -> (Int, Int) -> Bool

Função usada para calcular se um coordena vem depois de outra necessária para organizar as bombas.

Exemplos de utilização:

>>> coordCompare (0,0) (1,1)
False

checka :: String -> (Int, Int) -> Bool

Função que verifica se é possível ou não o jogador movimentar-se para uma determinada coordenada.

procLinha :: Int -> [String] -> String

Função que vai buscar a linha em que a função checka vai realizar a verificação.

moveRight :: (Int, Int) -> [String] -> (Int, Int)

Função que calcula as novas coordenadas se o jogador se poder mover ou não quando este se movimenta para a direita.

moveLeft :: (Int, Int) -> [String] -> (Int, Int)

Função que calcula as novas coordenadas se o jogador se poder mover ou não quando este se movimenta para a esquerda.

moveUP :: (Int, Int) -> [String] -> (Int, Int)

Função que calcula as novas coordenadas se o jogador se poder mover ou não quando este se movimenta para cima.

moveDown :: (Int, Int) -> [String] -> (Int, Int)

Função que calcula as novas coordenadas se o jogador se poder mover ou não quando este se movimenta para baixo.

checkCoord :: (Int, Int) -> [(Int, Int)] -> Bool

Função que verifica se um power up esta a ser usado.

takePUP :: [String] -> Int -> (Int, Int) -> [String]

Função que retira um power up do mapa quando este é apanhado.

testPUP :: String -> [Char] -> Bool

Função verifica se um jogador tem power ups.

Exemplos de utilização:

>>> testPUP "0 1 1 +"
True

Exemplos de utilização:

>>> testPUP "0 1 1"
False

finalPUP :: String -> String -> String -> String

Função que finaliza a string de um jogador com a informação dos power ups.

Exemplos de utilização:

>>> finalPUP "0 1 1" "+!" "+"
"0 1 1 +"

Exemplos de utilização:

>>> finalPUP "0 1 1 +" "+!" "!"
"0 1 1 +!"

verNumeB :: [String] -> Int -> Int

Função que calcula o numero de bombas que um jogador tem no mapa.

Exemplos de utilização:

>>> verNumeB ["* 1 1 0 1 4","* 1 2 0 1 4","* 3 1 1 1 4"] 0
2

firePOWER :: [String] -> Int -> Int

Função que calcula a potência das bombas de um jogador.

Exemplos de utilização:

>>> firePOWER ["0 1 1 !!","1 3 2","2 4 4"] 0
3

bombPOWER :: [String] -> Int -> Int

Função que calcula o número de bombas que um jogador pode colocar.

Exemplos de utilização:

>>> bombPOWER ["0 1 1","1 3 2 +","2 4 4"] 1
2

allowBomba :: [String] -> Int -> String

Função que dado um jogador verifica se este pode colocar uma bomba na sua posição.

stringBomba :: [String] -> Int -> (Int, Int) -> String

Função que origina um string com a informação de um bomba.

colocaBom :: [String] -> String -> [String]

Função que coloca uma bomba no mapa na sua posição correta.

powerUP :: [String] -> Int -> Char -> String

Função que verifica se o jogador apanhou um power up.

checkpowerU :: [String] -> Int -> String

Função que verifica se o jogador apanhou um power up quando se move para cima.

checkpowerL :: [String] -> Int -> String

Função que verifica se o jogador apanhou um power up quando se move para a esquerda.

checkpowerR :: [String] -> Int -> String

Função que verifica se o jogador apanhou um power up quando se move para a direita.

checkpowerD :: [String] -> Int -> String

Função que verifica se o jogador apanhou um power up quando se move para baixo.

newStr :: [String] -> Int -> Char -> String -> String

Função que altera as coordenadas de um jogador para as suas novas coordenadas.

move :: [String] -> Int -> Char -> [String]

A função move é a funçao principal e executa todos os comandos de movimentos e de bombas.

main :: IO ()