Main

Description

Com esta documentação pretendemos que se torne mais fácil a leitura e o entendimento do nosso código para fazer o jogo Bomberman. Aqui fica um pequeno resumo da lógica usada para fazer o jogo, e mais à frente abordamos todas as funções uma a uma. Nesta tarefa a função principal , avanca, divide-se em dois tipos de diferentes, uma responsável por executar a passagem do tempo sem o caracol ativado e outra responsável por executar a passagem do tempo com o caracol.

Documentation

type Dim = Int

Dim é a Dimensão do Mapa.

type Epicentro = (Int, Int)

Epicentro é um Ponto.

type Iplayer = String

String com a informação do Jogador.

type Idplayer = Int

Número de Jogador.

type Ipower = String

String com a informção de um powerup.

type Mapa = [String]

Mapa.

type Ibomb = String

String com a informação de uma bomba.

type World = (Mapa, [Ipower], [Ibomb], [Iplayer])

Mapa dividido em (Mapa,poderes,bombas,jogadores).

type Time = Int

Número de Ticks.

type Raio = Int

Raio de bomba.

type Lexplo = [(Int, Int)]

Lista de coordenadas.

avanca :: [String] -> Int -> [String]

A função avanca é a função responsável pela passagem do tempo.

Exemplos de utilização:

>>> avanca ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 1 2","0 2 1"] 200
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 1 1","0 2 1"]

>>> avanca ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 1 1","0 2 1"] 200
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]

semCaracol :: [String] -> [String]

A função semCaracol é a função responsavel por executar a passagem do tempo sem o caracol.

Exemplos de utilização:

>>> semCaracol ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 1 1","0 2 1"]
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]

comCaracol :: [String] -> Time -> [String]

A função comCaracol é a função responsavel por executar a passagem do tempo com o caracol.

Exemplos de utilização:

>>> comCaracol  ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 1 1","0 2 1"] 2
["#########","#       #","# #?#?# #","#  ?  ? #","#?### #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5"]

explodeTotal :: World -> [Ibomb] -> World

Função responsável por efetuar as alterações necessárias ao para resultantes das explosões.

Exemplos de utilização:

>>> explodeTotal (["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"],["+ 5 2","+ 3 3","! 5 5"],[],["0 2 1"]) ["* 1 1 0 1 1"]
(["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"],["+ 5 2","+ 3 3","! 5 5"],[],[])

splitMap :: [String] -> World

Função responsável por transformar a lista de Strings do mapa num tipo novo World que divide a lista de string em (Mapa,Poderes,Bombas,Jogadores) .

Exemplos de utilização:

>>> splitMap ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 1 1","0 2 1"]
(["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"],["+ 5 2","+ 3 3","! 5 5"],["* 1 1 0 1 1"],["0 2 1"])

groupMap :: World -> [String]

Função responsável por transformar um tipo World novamente em lista de strings .

Exemplos de utilização:

>>> groupMap (["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"],["+ 5 2","+ 3 3","! 5 5"],["* 1 1 0 1 1"],["0 2 1"])
["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 1 1","0 2 1"]

slowTime :: [Ibomb] -> [Ibomb]

Função responsável por reduzir o tempo das bombas.

Exemplos de utilização:

>>> slowTime ["* 1 1 0 2 5","* 1 4 1 2 3"]
["* 1 1 0 2 4","* 1 4 1 2 2"]

checkBombs :: [Ibomb] -> [Ibomb]

Função responsável por retirar as bombas com tick 0 , ou seja, que vao expludir.

Exemplos de utilização:

>>> checkBombs ["* 1 1 0 2 5","* 1 4 1 2 3"]
[]

checkBombs2 :: [Ibomb] -> [Ibomb]

Função responsável por retirar as bombas com tick diferente de 0 , ou seja, as que ficam no mapa.

Exemplos de utilização:

>>> checkBombs2 ["* 1 1 0 2 5","* 1 4 1 2 3"]
["* 1 1 0 2 5","* 1 4 1 2 3"]

coordPLayer :: Iplayer -> (Int, Int)

Função que dada um String respondes as coordenadas que estao na String.

Exemplos de utilização:

>>> coordPlayer "* 1 1 0 2 5"
(1,1)

coordMapa :: Mapa -> (Int, Int) -> Char

Função que dada umas coordenadas do Mapa responde o informação sobre o que esta nessa coordenada.

Exemplos de utilização:

>>> coordMapa ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########","+ 5 2","+ 3 3","! 5 5","* 1 1 0 1 1","0 2 1"] (2,1)
' '

killplayer :: [Iplayer] -> [(Int, Int)] -> [Iplayer]

Função que dada as Strings do jogadores e as coordenadas das explusões retira os jogadores que estao nessas coordenadas.

Exemplos de utilização:

>>> killplayer ["0 1 1","1 3 1"] [(1,1),(5,2),(8,9)]
["1 3 1"]

tiraj :: [Iplayer] -> (Int, Int) -> [Iplayer]

Função auxiliar da ´killplayer´ que dada um coordenada retira os jogadores nessa coordenada.

Exemplos de utilização:

>>> tiraj ["0 1 1","1 3 1"] (1,1)
["1 3 1"]

explode :: World -> Raio -> Epicentro -> World

Função que dado um Estado World, o raio da bomba e o seu centro faz a explusão da mesma.

Exemplos de utilização:

>>> explode (["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"],["+ 5 2","+ 3 3","! 5 5"],[],["0 2 1"]) 1 (1,1)
(["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"],["+ 5 2","+ 3 3","! 5 5"],[],[])

newBombs :: [Ibomb] -> [Epicentro] -> [Ibomb]

Função altera tick das bombas para 1 se ela estiver numa das coordenadas das explusões.

Exemplos de utilização:

>>> newBombs ["* 1 1 0 1 10","* 4 1 1 2 10"] [(1,1),(2,3)]
["* 1 1 0 1 1","* 4 1 1 2 10"]

showtime2 :: Ibomb -> [Epicentro] -> Ibomb

Função auxiliar da newBombs que altera o tick de uma bomba para 1 se ela estiver numa das coordenadas das explusões.

Exemplos de utilização:

>>> showtime2 "* 1 1 0 1 10" [(1,1),(2,3)]
"* 1 1 0 1 1"

checkraid :: Lexplo -> Mapa -> Lexplo -> Lexplo

Função que dada as todas as coordenadas possivel da explusão, o mapa e as coordenadas dos powerups devolve as coordenas de que não destroem ? no mapa.

Exemplos de utilização:

>>> checkraid [(3,2),(3,3)] ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"] [(5,2),(3,3),(5,5)]
[]

checkraid2 :: Lexplo -> Mapa -> Lexplo -> Lexplo

Função que dada as todas as coordenadas possivel da explusão, o mapa e as coordenadas dos powerups devolve as coordenas de que não destroem ? no mapa.

Exemplos de utilização:

>>> checkraid2 [(3,2),(3,3)] ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"] [(5,2),(3,3),(5,5)]
[(3,2)]

altotal :: Mapa -> [Epicentro] -> Mapa

Função que dada as todas as coordenadas a alterar no mapa, altera essas coordenadas.

Exemplos de utilização:

>>> altotal ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"] [(5,2),(3,3)]
["#########","#       #","# #?# # #","#     ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"]

altmapa :: Mapa -> Epicentro -> Char -> Mapa

Função que dada uma coordenada altera no mapa essa coordenada.

Exemplos de utilização:

>>> almapa ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"] (5,2) ' '
["#########","#       #","# #?# # #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"]

alteraL :: String -> Int -> Char -> String

Função que dado um Int altera a o elemento com essa posição na string.

Exemplos de utilização:

>>> alteraL "#########" 2 '!'
"##!######"

altotal2 :: Mapa -> [Epicentro] -> Mapa

Função igual à altotal2 mas neste caso utilizada para a expiral.

Exemplos de utilização:

>>> altotal2 ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"] [(5,2),(3,3)]
["#########","#       #","# #?### #","#  #  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"]

splitexplosion :: Epicentro -> Raio -> Mapa -> Lexplo -> Lexplo

Função que dado o centro, o raio, o mapa e os powerups produz uma lista com as coordenadas que alteram os powerups, os jogadores e as bombas.

Exemplos de utilização:

>>> splitexplosion (1,1) 1 ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"] [(5,2),(3,3)]
[(1,1),(1,2),(2,1)]

splitexplosion2 :: Epicentro -> Raio -> Mapa -> Lexplo -> Lexplo

Função que dado o centro, o raio, o mapa e os powerups produz uma lista com as coordenadas que alteram o mapa numeadamente os ? destruidos.

Exemplos de utilização:

>>> splitexplosion2 (1,1) 1 ["#########","#       #","# #?#?# #","#  ?  ? #","#?# # #?#","# ?  ?  #","# #?#?# #","#  ??   #","#########"] [(5,2),(3,3)]
[]

explosionU :: Epicentro -> Raio -> Int -> Lexplo

Função que dado o centro e o raio gera as coordenadas para cima com o centro.

Exemplos de utilização:

>>> explosionU (3,1)  2 0
[(3,1),(3,0)]

explosionD :: Epicentro -> Raio -> Int -> Lexplo

Função que dado o centro e o raio gera as coordenadas para baixo.

Exemplos de utilização:

>>> explosionD (3,1)  2 0
[(3,2),(3,3)]

explosionL :: Epicentro -> Raio -> Int -> Lexplo

Função que dado o centro e o raio gera as coordenadas para a esquerda.

Exemplos de utilização:

>>> explosionL (3,1)  2 0
[(2,1),(1,1)]

explosionR :: Epicentro -> Raio -> Int -> Lexplo

Função que dado o centro e o raio gera as coordenadas para cima com o centro

Exemplos de utilização:

>>> explosionR (3,1)  2 0
[(4,1),(5,1)]

map1 :: (a -> Int -> Int -> b) -> [a] -> Int -> Int -> [b]

Função auxiliar.

timeCheck :: Time -> Int -> Bool

Função que dado um tempo responde se o caracol esta ativo.

Exemplos de utilização:

>>> timeCheck 100 10
False

timeplay :: Time -> Int -> Int

Função que dado um tempo responde em que altura do caracol esta.

Exemplos de utilização:

>>> timeplay 100 10
6

criaCoord :: Dim -> (Int, Int) -> Int -> (Int, Int) -> [(Int, Int)]

Função uma dimensão e o ponto em que começa dá uma volta com as coordenadas utilizadas para a expiral.

Exemplos de utilização:

>>> criaCoord 7 (1,1) 0 (1,1)
[(1,1),(2,1),(3,1),(4,1),(5,1),(5,2),(5,3),(5,4),(5,5),(4,5),(3,5),(2,5),(1,5),(1,4),(1,3),(1,2)]

cria1 :: Dim -> (Int, Int) -> [(Int, Int)]

Função que dado dimensao e o ponto inical cria coordenadas para a direita.

Exemplos de utilização:

>>> cria1 3 (1,1)
[(1,1),(2,1),(3,1)]

cria2 :: Dim -> (Int, Int) -> [(Int, Int)]

Função que dado dimensao e o ponto inical cria coordenadas para baixo.

Exemplos de utilização:

>>> cria2 3 (1,1)
[(1,1),(1,2),(1,3)]

cria3 :: Dim -> (Int, Int) -> [(Int, Int)]

Função que dado dimensao e o ponto inical cria coordenadas para esquerda.

Exemplos de utilização:

>>> cria3 3 (1,1)
[(1,1),(0,1),(-1,1)]

cria4 :: Dim -> (Int, Int) -> [(Int, Int)]

Função que dado dimensao e o ponto inical cria coordenadas para cima.

Exemplos de utilização:

>>> cria4 3 (1,1)
[(1,1),(1,0),(1,-1)]

criaT :: Dim -> (Int, Int) -> Int -> [(Int, Int)]

Função que cria as coordenadas da espiral.

Exemplos de utilização:

>>> criaT 7 (1,1) 2
[(1,1),(2,1),(3,1),(4,1),(5,1),(5,2),(5,3),(5,4),(5,5),(4,5),(3,5),(2,5),(1,5),(1,4),(1,3),(1,2),(2,2),(3,2),(4,2),(4,3),(4,4),(3,4),(2,4),(2,3),(3,3)]

main :: IO ()

’main’: função que dá print ao mapa.